Create a schemas.py module
